Add multilanguage support
This commit relates to the issue #23 and provides an experimental
feature to index and search multiple language pages.

In order to create an index, that includes all the terms for each
available language of a page, the document id (on grav.index) is
now built as follows:

doc_id = page_lang:page_rawRoute

To search for a term a new parameter 'langs' is provided. The value
of 'langs' is a list of languages separated by colons, by default
either a fixed value provided on tntsearch.yaml or the current language.
After the results are found in the index, they are filtered according
if the prefix of the document id (page_lang) match any of the languages
requested by the 'langs' param.

Caveats:

Because the changes to add this feature are done on the plugin side
and not on the TNTSearch library, there is not really a multilanguage
index. For example, it is not possible to use a different stemmer to
index each term.

Closes #23